Update AuthorizationService.php
This looks better. The Assertions - as far as I understand - should only return the true case, if anything. If an assertion is `false`, `isGranted()` should still be able to return `true` if the permission is granted from the role. This change now fixes things, but we really need them unit tests :P #136